Moving regime common methods into more appropriate locations
