+++ regimes/ae/ae.go
@@ -6,7 +6,6 @@ import (
 	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/currency"
 	"github.com/invopop/gobl/i18n"
-	"github.com/invopop/gobl/regimes/common"
 	"github.com/invopop/gobl/tax"
 )
 
@@ -26,7 +25,7 @@ func New() *tax.RegimeDef {
 		},
 		TimeZone: "Asia/Dubai",
 		Tags: []*tax.TagSet{
-			common.InvoiceTags(),
+			bill.InvoiceTags(),
 		},
 		Scenarios: []*tax.ScenarioSet{
 			invoiceScenarios,
