+++ regimes/br/br.go
@@ -7,7 +7,6 @@ import (
 	"github.com/invopop/gobl/currency"
 	"github.com/invopop/gobl/i18n"
 	"github.com/invopop/gobl/org"
-	"github.com/invopop/gobl/regimes/common"
 	"github.com/invopop/gobl/tax"
 )
 
@@ -27,7 +26,7 @@ func New() *tax.RegimeDef {
 		TimeZone:  "America/Sao_Paulo",
 		Validator: Validate,
 		Tags: []*tax.TagSet{
-			common.InvoiceTags(),
+			bill.InvoiceTags(),
 		},
 		Categories: taxCategories,
 		Corrections: []*tax.CorrectionDefinition{
