+++ regimes/co/co.go
@@ -6,7 +6,6 @@ import (
 	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/i18n"
 	"github.com/invopop/gobl/org"
-	"github.com/invopop/gobl/regimes/common"
 	"github.com/invopop/gobl/tax"
 )
 
@@ -25,7 +24,7 @@ func New() *tax.RegimeDef {
 		},
 		TimeZone: "America/Bogota",
 		Tags: []*tax.TagSet{
-			common.InvoiceTags(),
+			bill.InvoiceTags(),
 		},
 		Validator:  Validate,
 		Normalizer: Normalize,
