+++ regimes/es/es.go
@@ -6,7 +6,6 @@ import (
 	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/currency"
 	"github.com/invopop/gobl/i18n"
-	"github.com/invopop/gobl/regimes/common"
 	"github.com/invopop/gobl/tax"
 )
 
@@ -60,7 +59,7 @@ func New() *tax.RegimeDef {
 		},
 		TimeZone: "Europe/Madrid",
 		Tags: []*tax.TagSet{
-			common.InvoiceTags().Merge(invoiceTags),
+			bill.InvoiceTags().Merge(invoiceTags),
 		},
 		Identities: identityDefinitions,
 		Categories: taxCategories,
