+++ regimes/fr/fr.go
@@ -7,7 +7,6 @@ import (
 	"github.com/invopop/gobl/currency"
 	"github.com/invopop/gobl/i18n"
 	"github.com/invopop/gobl/pkg/here"
-	"github.com/invopop/gobl/regimes/common"
 	"github.com/invopop/gobl/tax"
 )
 
@@ -42,7 +41,7 @@ func New() *tax.RegimeDef {
 		},
 		TimeZone: "Europe/Paris",
 		Tags: []*tax.TagSet{
-			common.InvoiceTags(),
+			bill.InvoiceTags(),
 		},
 		Scenarios: []*tax.ScenarioSet{
 			invoiceScenarios,
