+++ regimes/fr/tax_identity.go
@@ -7,7 +7,6 @@ import (
 	"strconv"
 
 	"github.com/invopop/gobl/cbc"
-	"github.com/invopop/gobl/regimes/common"
 	"github.com/invopop/gobl/tax"
 	"github.com/invopop/validation"
 )
@@ -85,10 +84,32 @@ func validateSIRENTaxCode(value interface{}) error {
 
 	base := str[:8]
 	chk := str[8:]
-	v := common.ComputeLuhnCheckDigit(base)
+	v := computeLuhnCheckDigit(base)
 	if chk != v {
 		return errors.New("checksum mismatch")
 	}
 
 	return nil
+}
+
+// TODO: refactor this into a shareable method.
+func computeLuhnCheckDigit(number string) string {
+	sum := 0
+	pos := 0
+
+	for i := len(number) - 1; i >= 0; i-- {
+		digit := int(number[i] - '0')
+
+		if pos%2 == 0 {
+			digit *= 2
+			if digit > 9 {
+				digit -= 9
+			}
+		}
+
+		sum += digit
+		pos++
+	}
+
+	return strconv.FormatInt(int64((10-(sum%10))%10), 10)
 }
