+++ regimes/pl/pl.go
@@ -6,7 +6,6 @@ import (
 	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/currency"
 	"github.com/invopop/gobl/i18n"
-	"github.com/invopop/gobl/regimes/common"
 	"github.com/invopop/gobl/tax"
 )
 
@@ -42,7 +41,7 @@ func New() *tax.RegimeDef {
 		PaymentMeansKeys: paymentMeansKeyDefinitions, // pay.go
 		Extensions:       extensionKeys,              // extensions.go
 		Tags: []*tax.TagSet{
-			common.InvoiceTags().Merge(invoiceTags),
+			bill.InvoiceTags().Merge(invoiceTags),
 		},
 		Scenarios:  scenarios, // scenarios.go
 		Validator:  Validate,
