+++ regimes/pt/pt.go
@@ -6,7 +6,6 @@ import (
 	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/currency"
 	"github.com/invopop/gobl/i18n"
-	"github.com/invopop/gobl/regimes/common"
 	"github.com/invopop/gobl/tax"
 )
 
@@ -47,7 +46,7 @@ func New() *tax.RegimeDef {
 		Validator:  Validate,
 		Normalizer: Normalize,
 		Tags: []*tax.TagSet{
-			common.InvoiceTags().Merge(invoiceTags),
+			bill.InvoiceTags().Merge(invoiceTags),
 		},
 		Corrections: []*tax.CorrectionDefinition{
 			{
