+++ regimes/us/us.go
@@ -6,7 +6,6 @@ import (
 	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/currency"
 	"github.com/invopop/gobl/i18n"
-	"github.com/invopop/gobl/regimes/common"
 	"github.com/invopop/gobl/tax"
 )
 
@@ -30,7 +29,7 @@ func New() *tax.RegimeDef {
 		TimeZone:  "America/Chicago", // Around the middle
 		Validator: Validate,
 		Tags: []*tax.TagSet{
-			common.InvoiceTags(),
+			bill.InvoiceTags(),
 		},
 		Categories: []*tax.CategoryDef{
 			//
